+++ src/Communication/Socket/WaitForDataInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace HeadlessChromium\Communication\Socket;
+
+interface WaitForDataInterface
+{
+    public function waitForData(float $maxSeconds): bool;
+}