+++ src/Dom/Node.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace HeadlessChromium\Dom;
 
+use HeadlessChromium\Clip;
 use HeadlessChromium\Communication\Message;
 use HeadlessChromium\Communication\Response;
 use HeadlessChromium\Exception\DomException;
+use HeadlessChromium\Exception\StaleElementException;
 use HeadlessChromium\Page;
 
 class Node
@@ -21,16 +23,37 @@ class Node
      */
     protected $nodeId;
 
+    /**
+     * @var bool
+     */
+    protected bool $isStale = false;
+
     public function __construct(Page $page, int $nodeId)
     {
         $this->page = $page;
         $this->nodeId = $nodeId;
+
+        $page->getSession()->on('method:DOM.documentUpdated', function (...$event): void {
+            $this->isStale = true;
+        });
+    }
+
+    public function getNodeId(): int
+    {
+        return $this->nodeId;
+    }
+
+    public function getNodeIdForRequest(): int
+    {
+        $this->prepareForRequest();
+
+        return $this->getNodeId();
     }
 
     public function getAttributes(): NodeAttributes
     {
         $message = new Message('DOM.getAttributes', [
-            'nodeId' => $this->nodeId,
+            'nodeId' => $this->getNodeIdForRequest(),
         ]);
         $response = $this->page->getSession()->sendMessageSync($message);
 
@@ -44,7 +67,7 @@ class Node
     public function setAttributeValue(string $name, string $value): void
     {
         $message = new Message('DOM.setAttributeValue', [
-            'nodeId' => $this->nodeId,
+            'nodeId' => $this->getNodeIdForRequest(),
             'name' => $name,
             'value' => $value,
         ]);
@@ -56,7 +79,7 @@ class Node
     public function querySelector(string $selector): ?self
     {
         $message = new Message('DOM.querySelector', [
-            'nodeId' => $this->nodeId,
+            'nodeId' => $this->getNodeIdForRequest(),
             'selector' => $selector,
         ]);
         $response = $this->page->getSession()->sendMessageSync($message);
@@ -74,7 +97,7 @@ class Node
     public function querySelectorAll(string $selector): array
     {
         $message = new Message('DOM.querySelectorAll', [
-            'nodeId' => $this->nodeId,
+            'nodeId' => $this->getNodeIdForRequest(),
             'selector' => $selector,
         ]);
         $response = $this->page->getSession()->sendMessageSync($message);
@@ -93,7 +116,7 @@ class Node
     public function focus(): void
     {
         $message = new Message('DOM.focus', [
-            'nodeId' => $this->nodeId,
+            'nodeId' => $this->getNodeIdForRequest(),
         ]);
         $response = $this->page->getSession()->sendMessageSync($message);
 
@@ -108,7 +131,7 @@ class Node
     public function getPosition(): ?NodePosition
     {
         $message = new Message('DOM.getBoxModel', [
-            'nodeId' => $this->nodeId,
+            'nodeId' => $this->getNodeIdForRequest(),
         ]);
         $response = $this->page->getSession()->sendMessageSync($message);
 
@@ -131,7 +154,7 @@ class Node
     public function getHTML(): string
     {
         $message = new Message('DOM.getOuterHTML', [
-            'nodeId' => $this->nodeId,
+            'nodeId' => $this->getNodeIdForRequest(),
         ]);
         $response = $this->page->getSession()->sendMessageSync($message);
 
@@ -143,7 +166,7 @@ class Node
     public function setHTML(string $outerHTML): void
     {
         $message = new Message('DOM.setOuterHTML', [
-            'nodeId' => $this->nodeId,
+            'nodeId' => $this->getNodeIdForRequest(),
             'outerHTML' => $outerHTML,
         ]);
         $response = $this->page->getSession()->sendMessageSync($message);
@@ -159,7 +182,7 @@ class Node
     public function scrollIntoView(): void
     {
         $message = new Message('DOM.scrollIntoViewIfNeeded', [
-            'nodeId' => $this->nodeId,
+            'nodeId' => $this->getNodeIdForRequest(),
         ]);
         $response = $this->page->getSession()->sendMessageSync($message);
 
@@ -177,7 +200,7 @@ class Node
         $this->scrollIntoView();
         $position = $this->getPosition();
         $this->page->mouse()
-            ->move($position->getCenterX(), $position->getCenterY())
+            ->move((int) $position->getCenterX(), (int) $position->getCenterY())
             ->click();
     }
 
@@ -198,7 +221,7 @@ class Node
     {
         $message = new Message('DOM.setFileInputFiles', [
             'files' => $filePaths,
-            'nodeId' => $this->nodeId,
+            'nodeId' => $this->getNodeIdForRequest(),
         ]);
         $response = $this->page->getSession()->sendMessageSync($message);
 
@@ -212,6 +235,33 @@ class Node
     {
         if (!$response->isSuccessful()) {
             throw new DOMException($response->getErrorMessage());
+        }
+    }
+
+    public function getClip(): ?Clip
+    {
+        $position = $this->getPosition();
+
+        if (!$position) {
+            return null;
+        }
+
+        return new Clip(
+            $position->getX(),
+            $position->getY(),
+            $position->getWidth(),
+            $position->getHeight(),
+        );
+    }
+
+    protected function prepareForRequest(): void
+    {
+        $this->page->assertNotClosed();
+
+        $this->page->getSession()->getConnection()->processAllEvents();
+
+        if ($this->isStale) {
+            throw new StaleElementException();
         }
     }
 }