+++ src/Exception/CommunicationException/CantSyncEventsException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace HeadlessChromium\Exception\CommunicationException;
+
+use HeadlessChromium\Exception\CommunicationException;
+
+class CantSyncEventsException extends CommunicationException
+{
+}