+++ src/Exception/StaleElementException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace HeadlessChromium\Exception;
+
+class StaleElementException extends DomException
+{
+}